adding ui to scene. adding right click to hurt yourself to test health bar. added a default ui background that we can just drag and drop where we please and change them all consistently

+++ Assets/Scripts/Entities/Damageable.cs
@@ -23,7 +23,7 @@ public class Damageable : MonoBehaviour
 
         health -= amount;
 
-        HealthUpdatedEvent.Invoke(health / maxHealth);
+        HealthUpdatedEvent.Invoke((float)health / maxHealth);
 
         if (health <= 0)
             DoDie();
